move metadata merging code to seperate files
